+++ packages/extensions/core/src/lib/configuration/logger.ts
@@ -180,6 +180,7 @@ export class AutorestCoreLogger {
               });
             }
           } catch (e) {
+            console.error("E", e);
             if (!shouldComplain) {
               shouldComplain = true;
             }
