+++ packages/libs/common/src/literate-yaml/markdown-parser.ts
@@ -23,7 +23,6 @@ export async function parseCodeBlocksFromMarkdown(
 
     const data = codeBlock.literal || "";
     const mappings = getSourceMapForCodeBlock(hConfigFile.key, codeBlock);
-
     const hCodeBlock = await sink.WriteData(codeBlockKey, data, hConfigFile.identity, undefined, mappings, [
       hConfigFile,
     ]);
