+++ packages/libs/datastore/src/data-store/data-store.ts
@@ -3,12 +3,11 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { OperationCanceledException, Delay, Lazy } from "@azure-tools/tasks";
-import { ReadUri, ResolveUri, ParentFolderUri } from "@azure-tools/uri";
-import { MappedPosition, Position, RawSourceMap, SourceMapGenerator } from "source-map";
+import { OperationCanceledException, Lazy, LazyPromise } from "@azure-tools/tasks";
+import { resolveUri } from "@azure-tools/uri";
+import { MappingItem, Position, RawSourceMap, SourceMapConsumer, SourceMapGenerator } from "source-map";
 import { CancellationToken } from "../cancellation";
 import { IFileSystem } from "../file-system/file-system";
-import { FastStringify, ParseNode, ParseToAst as parseAst, YAMLNode, parseYaml } from "../yaml";
 import { BlameTree } from "../source-map/blaming";
 import { Compile, CompilePosition, Mapping, SmartPosition } from "../source-map/source-map";
 import { promises as fs } from "fs";
@@ -16,8 +15,12 @@ import { tmpdir } from "os";
 import { join } from "path";
 
 import { createHash } from "crypto";
-import { LineIndices } from "../main";
-import { logger } from "../logger";
+import { DataSource } from "./data-source";
+import { ReadThroughDataSource } from "./read-through-data-source";
+import { Data, DataHandle, Metadata } from "./data-handle";
+import { DataSink } from "./data-sink";
+import { LineIndices } from "../parsing/text-utility";
+
 const md5 = (content: any) => (content ? createHash("md5").update(JSON.stringify(content)).digest("hex") : null);
 
 const FALLBACK_DEFAULT_OUTPUT_ARTIFACT = "";
@@ -26,153 +29,8 @@ const FALLBACK_DEFAULT_OUTPUT_ARTIFACT = "";
  * Data model section (not exposed)
  ********************************************/
 
-export interface Metadata {
-  lineIndices: Lazy<Array<number>>;
-
-  // inputSourceMap: LazyPromise<RawSourceMap>;
-  // sourceMap: LazyPromise<RawSourceMap>;
-  // sourceMapEachMappingByLine: LazyPromise<Array<Array<MappingItem>>>;
-}
-
-export interface Data {
-  name: string;
-  artifactType: string;
-  metadata: Metadata;
-  identity: Array<string>;
-
-  writeToDisk?: Promise<void>;
-  cached?: string;
-  cachedAst?: YAMLNode;
-  cachedObject?: any;
-  accessed?: boolean;
-}
-
 interface Store {
   [uri: string]: Data;
-}
-
-/********************************************
- * Central data controller
- * - one stop for creating data
- * - ensures WRITE ONCE model
- ********************************************/
-
-export interface PipeState {
-  skipping?: boolean;
-  excludeFromCache?: boolean;
-}
-
-export function mergePipeStates(result: PipeState, ...pipeStates: Array<PipeState>) {
-  for (const each of pipeStates) {
-    result.skipping === undefined ? each.skipping : result.skipping && each.skipping;
-    result.excludeFromCache === undefined ? each.excludeFromCache : result.excludeFromCache || each.excludeFromCache;
-  }
-  return result;
-}
-
-export abstract class DataSource {
-  public pipeState: PipeState;
-  public abstract Enum(): Promise<Array<string>>;
-  public abstract Read(uri: string): Promise<DataHandle | null>;
-
-  constructor() {
-    this.pipeState = {};
-  }
-
-  get skip(): boolean {
-    return !!this.pipeState.skipping;
-  }
-  get cachable(): boolean {
-    return !this.pipeState.excludeFromCache;
-  }
-  set cachable(value: boolean) {
-    this.pipeState.excludeFromCache = !value;
-  }
-
-  public async ReadStrict(uri: string): Promise<DataHandle> {
-    const result = await this.Read(uri);
-    if (result === null) {
-      throw new Error(`Could not read '${uri}'.`);
-    }
-    return result;
-  }
-}
-
-export class QuickDataSource extends DataSource {
-  public constructor(private handles: Array<DataHandle>, pipeState?: PipeState) {
-    super();
-    this.pipeState = pipeState || {};
-  }
-
-  public async Enum(): Promise<Array<string>> {
-    return this.pipeState.skipping ? new Array<string>() : this.handles.map((x) => x.key);
-  }
-
-  public async Read(key: string): Promise<DataHandle | null> {
-    if (this.pipeState.skipping) {
-      return null;
-    }
-    const data = this.handles.filter((x) => x.key === key)[0];
-    return data || null;
-  }
-}
-
-class ReadThroughDataSource extends DataSource {
-  private uris: Array<string> = [];
-  private cache: { [uri: string]: Promise<DataHandle | null> } = {};
-
-  constructor(private store: DataStore, private fs: IFileSystem) {
-    super();
-  }
-
-  get cachable(): boolean {
-    // filesystem based data source can't cache
-    return false;
-  }
-
-  public async Read(uri: string): Promise<DataHandle | null> {
-    // sync cache (inner stuff is racey!)
-    if (!this.cache[uri]) {
-      this.cache[uri] = (async () => {
-        // probe data store
-        try {
-          const existingData = await this.store.Read(uri);
-          this.uris.push(uri);
-          return existingData;
-        } catch (e) {
-          // ignore .
-        }
-
-        // populate cache
-        let data: string | null = null;
-        try {
-          data = (await this.fs.read(uri)) || (await ReadUri(uri));
-          if (data) {
-            const parent = ParentFolderUri(uri) || "";
-            // hack to let $(this-folder) resolve to the location...
-            data = data.replace(/\$\(this-folder\)\/*/g, parent);
-          }
-        } catch (e) {
-          logger.error("Unexpected error trying to read file", e);
-        } finally {
-          if (!data) {
-            // eslint-disable-next-line no-unsafe-finally
-            return null;
-          }
-        }
-        const readHandle = await this.store.WriteData(uri, data, "input-file", [uri]);
-
-        this.uris.push(uri);
-        return readHandle;
-      })();
-    }
-
-    return this.cache[uri];
-  }
-
-  public async Enum(): Promise<Array<string>> {
-    return this.uris;
-  }
 }
 
 export class DataStore {
@@ -252,50 +110,50 @@ export class DataStore {
 
     // metadata.artifactType = artifact;
 
-    // metadata.sourceMap = new LazyPromise(async () => {
-    //   if (!sourceMapFactory) {
-    //     return new SourceMapGenerator().toJSON();
-    //   }
-    //   const sourceMap = await sourceMapFactory(result);
+    metadata.sourceMap = new LazyPromise(async () => {
+      if (!sourceMapFactory) {
+        return new SourceMapGenerator().toJSON();
+      }
+      const sourceMap = await sourceMapFactory(result);
 
-    //   // validate
-    //   const inputFiles = sourceMap.sources.concat(sourceMap.file);
-    //   for (const inputFile of inputFiles) {
-    //     if (!this.store[inputFile]) {
-    //       throw new Error(`Source map of '${uri}' references '${inputFile}' which does not exist`);
-    //     }
-    //   }
+      // validate
+      const inputFiles = sourceMap.sources.concat(sourceMap.file);
+      for (const inputFile of inputFiles) {
+        if (!this.store[inputFile]) {
+          throw new Error(`Source map of '${uri}' references '${inputFile}' which does not exist`);
+        }
+      }
 
-    //   return sourceMap;
-    // });
+      return sourceMap;
+    });
 
-    // metadata.sourceMapEachMappingByLine = new LazyPromise<Array<Array<MappingItem>>>(async () => {
-    //   const result: Array<Array<MappingItem>> = [];
+    metadata.sourceMapEachMappingByLine = new LazyPromise<Array<Array<MappingItem>>>(async () => {
+      const result: Array<Array<MappingItem>> = [];
 
-    //   const sourceMapConsumer = new SourceMapConsumer(await metadata.sourceMap);
+      const sourceMapConsumer = new SourceMapConsumer(await metadata.sourceMap);
 
-    //   // does NOT support multiple sources :(
-    //   // `singleResult` has null-properties if there is no original
+      // does NOT support multiple sources :(
+      // `singleResult` has null-properties if there is no original
 
-    //   // get coinciding sources
-    //   sourceMapConsumer.eachMapping(mapping => {
-    //     while (result.length <= mapping.generatedLine) {
-    //       result.push([]);
-    //     }
-    //     result[mapping.generatedLine].push(mapping);
-    //   });
+      // get coinciding sources
+      sourceMapConsumer.eachMapping((mapping) => {
+        while (result.length <= mapping.generatedLine) {
+          result.push([]);
+        }
+        result[mapping.generatedLine].push(mapping);
+      });
 
-    //   return result;
-    // });
+      return result;
+    });
 
-    // metadata.inputSourceMap = new LazyPromise(() => this.CreateInputSourceMapFor(uri));
+    metadata.inputSourceMap = new LazyPromise(() => this.CreateInputSourceMapFor(uri));
     metadata.lineIndices = new Lazy<Array<number>>(() => LineIndices(data));
 
     return result;
   }
 
   private createUri(description: string): string {
-    return ResolveUri(this.BaseUri, `${this.uid++}?${encodeURIComponent(description)}`);
+    return resolveUri(this.BaseUri, `${this.uid++}?${encodeURIComponent(description)}`);
   }
 
   public getDataSink(defaultArtifact: string = FALLBACK_DEFAULT_OUTPUT_ARTIFACT): DataSink {
@@ -319,7 +177,7 @@ export class DataStore {
   }
 
   public async Read(uri: string): Promise<DataHandle> {
-    uri = ResolveUri(this.BaseUri, uri);
+    uri = resolveUri(this.BaseUri, uri);
     const data = this.store[uri];
     if (!data) {
       throw new Error(`Could not read '${uri}'.`);
@@ -337,7 +195,7 @@ export class DataStore {
       name: `blameRoot (${JSON.stringify(position)})`,
     });
   }
-  /* DISABLING SOURCE MAP SUPPORT
+
   private async CreateInputSourceMapFor(absoluteUri: string): Promise<RawSourceMap> {
     const data = this.ReadStrictSync(absoluteUri);
 
@@ -345,7 +203,9 @@ export class DataStore {
     const targetPositions: Array<SmartPosition> = [];
     const metadata = data.metadata;
     const sourceMapConsumer = new SourceMapConsumer(await metadata.sourceMap);
-    sourceMapConsumer.eachMapping(m => targetPositions.push(<Position>{ column: m.generatedColumn, line: m.generatedLine }));
+    sourceMapConsumer.eachMapping((m) =>
+      targetPositions.push(<Position>{ column: m.generatedColumn, line: m.generatedLine }),
+    );
 
     // collect blame
     const mappings: Array<Mapping> = [];
@@ -357,227 +217,12 @@ export class DataStore {
           name: inputPosition.name,
           source: this.ReadStrictSync(inputPosition.source).Description, // friendly name
           generated: blameTree.node,
-          original: inputPosition
+          original: inputPosition,
         });
       }
     }
     const sourceMapGenerator = new SourceMapGenerator({ file: absoluteUri });
-    Compile(mappings, sourceMapGenerator);
+    await Compile(mappings, sourceMapGenerator);
     return sourceMapGenerator.toJSON();
-  }
-  */
-}
-
-/********************************************
- * Data handles
- * - provide well-defined access to specific data
- * - provide convenience methods
- ********************************************/
-
-export class DataSink {
-  constructor(
-    private write: (
-      description: string,
-      rawData: string,
-      artifact: string | undefined,
-      identity: Array<string>,
-      metadataFactory: (readHandle: DataHandle) => Promise<RawSourceMap>,
-    ) => Promise<DataHandle>,
-    private forward: (description: string, input: DataHandle) => Promise<DataHandle>,
-  ) {}
-
-  public async WriteDataWithSourceMap(
-    description: string,
-    data: string,
-    artifact: string | undefined,
-    identity: Array<string>,
-    sourceMapFactory: (readHandle: DataHandle) => Promise<RawSourceMap>,
-  ): Promise<DataHandle> {
-    return this.write(description, data, artifact, identity, sourceMapFactory);
-  }
-
-  public async WriteData(
-    description: string,
-    data: string,
-    identity: Array<string>,
-    artifact?: string,
-    mappings: Array<Mapping> = [],
-    mappingSources: Array<DataHandle> = [],
-  ): Promise<DataHandle> {
-    return this.WriteDataWithSourceMap(description, data, artifact, identity, async (readHandle) => {
-      const sourceMapGenerator = new SourceMapGenerator({ file: readHandle.key });
-      await Compile(mappings, sourceMapGenerator, mappingSources.concat(readHandle));
-      return sourceMapGenerator.toJSON();
-    });
-  }
-
-  public WriteObject<T>(
-    description: string,
-    obj: T,
-    identity: Array<string>,
-    artifact?: string,
-    mappings: Array<Mapping> = [],
-    mappingSources: Array<DataHandle> = [],
-  ): Promise<DataHandle> {
-    return this.WriteData(description, FastStringify(obj), identity, artifact, mappings, mappingSources);
-  }
-
-  public Forward(description: string, input: DataHandle): Promise<DataHandle> {
-    return this.forward(description, input);
-  }
-}
-
-export class DataHandle {
-  /**
-   * @param autoUnload If the data unhandle should automatically unload files after they are not used for a while.
-   */
-  constructor(public readonly key: string, private item: Data, private autoUnload = true) {
-    // start the clock once this has been created.
-    // this ensures that the data cache will be flushed if not
-    // used in a reasonable amount of time
-    this.onTimer();
-  }
-
-  public async serialize() {
-    this.item.name;
-    return JSON.stringify({
-      key: this.Description,
-      artifactType: this.item.artifactType,
-      identity: this.item.identity,
-      name: this.item.name,
-      content: await this.ReadData(true),
-    });
-  }
-
-  private onTimer() {
-    this.checkIfNeedToUnload().catch((e) => {
-      // eslint-disable-next-line no-console
-      console.error("Error while verifing DataHandle cache status", e);
-    });
-  }
-
-  private async checkIfNeedToUnload() {
-    if (!this.autoUnload) {
-      return;
-    }
-
-    await Delay(3000);
-    if (this.item.accessed) {
-      // it's been cached. start the timer!
-      this.onTimer();
-      // clear the accessed flag before we go.
-      this.item.accessed = false;
-      return;
-    }
-    // wasn't actually used since the delay. let's dump it.
-    // console.log(`flushing ${this.item.name}`);
-    // wait to make sure it's finished writing to disk tho'
-    // await this.item.writingToDisk;
-    if (!this.item.writeToDisk) {
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      this.item.writeToDisk = fs.writeFile(this.item.name, this.item.cached!);
-    }
-    // clear the caches.
-    this.item.cached = undefined;
-    this.item.cachedObject = undefined;
-    this.item.cachedAst = undefined;
-  }
-
-  public get originalDirectory() {
-    const id = this.identity[0];
-    return id.substring(0, id.lastIndexOf("/"));
-  }
-
-  public get originalFullPath() {
-    return this.identity[0];
-  }
-
-  public get identity() {
-    return this.item.identity;
-  }
-
-  public async ReadData(nocache = false): Promise<string> {
-    if (!nocache) {
-      // we're going to use the data, so let's not let it expire.
-      this.item.accessed = true;
-    }
-    // if it's not cached, load it from disk.
-    if (this.item.cached === undefined) {
-      // make sure the write-to-disk is finished.
-      await this.item.writeToDisk;
-
-      if (nocache) {
-        return await fs.readFile(this.item.name, "utf8");
-      } else {
-        this.item.cached = await fs.readFile(this.item.name, "utf8");
-
-        // start the timer again.
-        this.onTimer();
-      }
-    }
-
-    return this.item.cached;
-  }
-
-  public async ReadObjectFast<T>(): Promise<T> {
-    // we're going to use the data, so let's not let it expire.
-    this.item.accessed = true;
-
-    return this.item.cachedObject || (this.item.cachedObject = parseYaml(await this.ReadData()));
-  }
-
-  public async ReadObject<T>(): Promise<T> {
-    // we're going to use the data, so let's not let it expire.
-    this.item.accessed = true;
-
-    // return the cached object, or get it, then return it.
-    return this.item.cachedObject || (this.item.cachedObject = ParseNode<T>(await this.ReadYamlAst()));
-  }
-
-  public async ReadYamlAst(): Promise<YAMLNode> {
-    // we're going to use the data, so let's not let it expire.
-    this.item.accessed = true;
-    // return the cachedAst or get it, then return it.
-    return this.item.cachedAst || (this.item.cachedAst = parseAst(await this.ReadData()));
-  }
-
-  public get metadata(): Metadata {
-    return this.item.metadata;
-  }
-
-  public get artifactType(): string {
-    return this.item.artifactType;
-  }
-
-  public get Description(): string {
-    return decodeURIComponent(this.key.split("?").reverse()[0]);
-  }
-
-  public async IsObject(): Promise<boolean> {
-    try {
-      await this.ReadObject();
-      return true;
-    } catch (e) {
-      return false;
-    }
-  }
-
-  public async Blame(position: Position): Promise<Array<MappedPosition>> {
-    return [];
-    /* DISABLING SOURCE MAP SUPPORT
-    const metadata = this.metadata;
-    const sameLineResults = ((await metadata.sourceMapEachMappingByLine)[position.line] || [])
-      .filter(mapping => mapping.generatedColumn <= position.column);
-    const maxColumn = sameLineResults.reduce((c, m) => Math.max(c, m.generatedColumn), 0);
-    const columnDelta = position.column - maxColumn;
-    return sameLineResults.filter(m => m.generatedColumn === maxColumn).map(m => {
-      return {
-        column: m.originalColumn + columnDelta,
-        line: m.originalLine,
-        name: m.name,
-        source: m.source
-      };
-    });
-    */
   }
 }
