+++ packages/libs/datastore/src/parsing/yaml.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { DataHandle } from "../data-store/data-store";
+import { DataHandle } from "../data-store";
 import { JsonPath, JsonPathComponent, stringify } from "../json-path/json-path";
 import { EnhancedPosition } from "../source-map/source-map";
 import {
