+++ packages/libs/datastore/src/source-map/source-map.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Position, SourceMapGenerator } from "source-map";
-import { DataHandle } from "../data-store/data-store";
+import { DataHandle } from "../data-store";
 import { JsonPath, stringify } from "../json-path/json-path";
 import { IndexToPosition } from "../parsing/text-utility";
 import * as yaml from "../parsing/yaml";
